+++ Core/include/EventDatabaseInterface.hxx
@@ -18,8 +18,8 @@ public:
 	}
 
 	virtual void addEvent( EventInterface* event )=0;
-	virtual int getExpectedID( EventInterface* event ) const=0; ///< Powoduje znaczną niestabilność wiązanie events - marks
-	virtual EventInterface* getEventByID( int ID ) const=0;
+	virtual qint64 getExpectedID( EventInterface* event ) const=0; ///< Powoduje znaczną niestabilność wiązanie events - marks
+	virtual EventInterface* getEventByID( qint64 ID ) const=0;
 };
 
 #endif /* DATABESAINTERFACE_HXX_ */
